pyproject.toml
[project]
name = "capfit"
version = "0.1.15"
description = "Split long chat screenshots into column-sized chunks and build 2-column PDFs."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Name", email = "you@example.com" }]
license = { text = "MIT" }
dependencies = [
  "pillow>=10.2.0",
  "numpy>=1.26.0",
  "typer[all]>=0.12.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.23.0",
  "jinja2>=3.1.2",
  "python-multipart>=0.0.9",
]

[project.scripts]
capfit = "cli.app:main"
capfit-web = "web.server:main"
capfit-desktop = "desktop.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["shared*", "web*", "desktop*", "cli*"]

[tool.setuptools.package-data]
web = [
  "templates/*.html",
  "static/*",
]
desktop = [
  "assets/*",
]
